+++ src/org/petctviewer/orthanc/anonymize/controllers/Controller_Csv_Btn.java
@@ -2,10 +2,16 @@ package org.petctviewer.orthanc.anonymize.controllers;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.io.FileNotFoundException;
+import java.io.File;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.prefs.Preferences;
 
+import javax.swing.JFileChooser;
+
+import org.petctviewer.orthanc.Orthanc_Tools;
 import org.petctviewer.orthanc.anonymize.CSV;
-import org.petctviewer.orthanc.anonymize.QueryOrthancData;
 import org.petctviewer.orthanc.anonymize.TableExportStudiesModel;
 import org.petctviewer.orthanc.anonymize.VueAnon;
 import org.petctviewer.orthanc.anonymize.datastorage.Study2;
@@ -14,11 +20,11 @@ import org.petctviewer.orthanc.anonymize.datastorage.Study_Anonymized;
 public class Controller_Csv_Btn implements ActionListener {
 	
 	TableExportStudiesModel modeleExportStudies;
-	QueryOrthancData queryOrthanc;
+	VueAnon vueAnon;
 
-	public Controller_Csv_Btn(TableExportStudiesModel modeleExportStudies, QueryOrthancData queryOrthanc) {
+	public Controller_Csv_Btn(TableExportStudiesModel modeleExportStudies, VueAnon vueAnon) {
 		this.modeleExportStudies=modeleExportStudies;
-		this.queryOrthanc=queryOrthanc;
+		this.vueAnon=vueAnon;
 	}
 	@Override
 	public void actionPerformed(ActionEvent arg0) {
@@ -26,29 +32,39 @@ public class Controller_Csv_Btn implements ActionListener {
 		VueAnon.jprefer.put("reportType", "CSV");
 		CSV csv = new CSV();
 		if(!modeleExportStudies.getOrthancIds().isEmpty()){
+			
 			for(Study_Anonymized anonymizedObj : modeleExportStudies.getAnonymizedObject()){
 				Study2 anonymizedstudy=anonymizedObj.getAnonymizedStudy();
 				Study2 originStudy=anonymizedObj.getOriginalStudy();
 				
-				anonymizedstudy.storeStudyStatistics(queryOrthanc);
+				anonymizedstudy.storeStudyStatistics(vueAnon.getOrthancQuery());
 				
 				csv.addStudy(originStudy.getPatientName(), originStudy.getPatientID(), 
 						anonymizedstudy.getPatientName(), anonymizedstudy.getPatientID(),
 						originStudy.getDate(), originStudy.getStudyDescription(), 
 						anonymizedstudy.getStudyDescription(), anonymizedstudy.getStatNbSeries(), 
 						anonymizedstudy.getStatNbInstance(), anonymizedstudy.getMbSize(), anonymizedstudy.getStudyInstanceUid());
-					
-		
+
 			}
-			try {
-				csv.genCSV();
-			} catch (FileNotFoundException e1) {
-				e1.printStackTrace();
+			
+			
+			Preferences jpreferAnon = VueAnon.jprefer;
+			DateFormat df = new SimpleDateFormat("MM_dd_yyyy_HHmmss");
+			
+			JFileChooser chooser = new JFileChooser();
+			chooser.setCurrentDirectory(new File(jpreferAnon.get("csvLocation", System.getProperty("user.dir"))));
+			chooser.setSelectedFile(new File(df.format(new Date()) + ".csv"));
+			chooser.setDialogTitle("Export csv to...");
+			chooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
+			chooser.setAcceptAllFileFilterUsed(false);
+
+			if (chooser.showSaveDialog(vueAnon) == JFileChooser.APPROVE_OPTION) {
+				File fichier=chooser.getSelectedFile();
+				jpreferAnon.put("csvLocation", chooser.getSelectedFile().toPath().toString() );
+				Orthanc_Tools.writeCSV(csv.getCsv(), fichier);
 			}
 			
 		}
-		
-	
 		
 	}
