+++ src/org/petctviewer/orthanc/anonymize/datastorage/Tags.java
@@ -15,7 +15,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-package org.petctviewer.orthanc.anonymize;
+package org.petctviewer.orthanc.anonymize.datastorage;
 
 /**
  * Store Choice for each Tag for anonymization query construction
